+++ test/index_ensure_test.go
@@ -501,3 +501,68 @@ func TestEnsureZKDIndexWithOptions(t *testing.T) {
 	err = idx.Remove(nil)
 	require.NoError(t, err)
 }
+
+// TestEnsureInvertedIndex creates a collection with an inverted index
+func TestEnsureInvertedIndex(t *testing.T) {
+	ctx := context.Background()
+
+	c := createClientFromEnv(t, true)
+	EnsureVersion(t, ctx, c).CheckVersion(MinimumVersion("3.10.0"))
+
+	db := ensureDatabase(ctx, c, "index_test", nil, t)
+	col := ensureCollection(ctx, db, fmt.Sprintf("inverted_index_opt_test"), nil, t)
+
+	opt := driver.InvertedIndexOptions{
+		Name: "inverted-opt",
+		PrimarySort: driver.InvertedIndexPrimarySort{
+			Fields: []driver.ArangoSearchPrimarySortEntry{
+				{Field: "test1", Ascending: newBool(true)},
+				{Field: "test2", Ascending: newBool(false)},
+			},
+			Compression: driver.PrimarySortCompressionLz4,
+		},
+		Features:     []string{},
+		StoredValues: []driver.StoredValue{},
+		Fields: []driver.InvertedIndexField{
+			{Name: "field1", Features: []string{"frequency"}, Nested: nil},
+			{Name: "field2", Features: []string{"position"}, TrackListPositions: false, Nested: []driver.InvertedIndexField{
+				{
+					Name: "some-nested-field",
+					Nested: []driver.InvertedIndexField{
+						{Name: "test"},
+						{Name: "bas", Nested: []driver.InvertedIndexField{
+							{Name: "a", Features: nil},
+						}},
+						{Name: "kas", Nested: []driver.InvertedIndexField{
+							{Name: "b", TrackListPositions: true},
+							{Name: "c"},
+						}},
+					},
+				},
+			}},
+		},
+	}
+
+	idx, created, err := col.EnsureInvertedIndex(ctx, &opt)
+	require.NoError(t, err)
+	require.True(t, created)
+
+	opt.IsNewlyCreated = true
+	opt.Analyzer = driver.ArangoSearchAnalyzerTypeIdentity // default value for analyzer
+
+	requireIdxEquality := func(invertedIdx driver.Index) {
+		require.Equal(t, driver.InvertedIndex, idx.Type())
+		require.Equal(t, opt.Name, idx.UserName())
+		require.Equal(t, opt, idx.InvertedIndexOptions())
+	}
+	requireIdxEquality(idx)
+
+	indexes, err := col.Indexes(ctx)
+	require.NoError(t, err)
+	require.NotEmpty(t, indexes)
+
+	requireIdxEquality(indexes[0])
+
+	err = idx.Remove(nil)
+	require.NoError(t, err)
+}