+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add support for optimizer rules (AQL query)
 - New `LegacyPolygons` parameter for Geo Indexes
 - New parameters (`cacheEnabled` and `storedValues`) for Persistent Indexes
+- Add support for Inverted index
 
 ## [1.3.3](https://github.com/arangodb/go-driver/tree/v1.3.3) (2022-07-27)
 - Fix `lastValue` field type